+++ Echo/src/utopia/echo/model/response/BufferedResponse.scala
@@ -1,7 +1,5 @@
 package utopia.echo.model.response
 
-import utopia.annex.model.manifest.HasSchrodingerState
-import utopia.annex.model.manifest.SchrodingerState.{Alive, Final}
 import utopia.flow.time.Now
 
 import java.time.Instant
@@ -12,18 +10,19 @@ object BufferedResponse
 	
 	/**
 	 * @param text Text contents of this response
+	 * @param thoughts The reflective / reasoning content produced by the LLM before the final answer.
 	 * @param tokenUsage Statistics about token usage
-	 * @param state Whether this response was a success or a failure
 	 * @param created Time when this (version of this) response was created
 	 * @return A new response
 	 */
-	def apply(text: String, tokenUsage: TokenUsage, state: Final = Alive, created: Instant = Now): BufferedResponse =
-		_BufferedResponse(text, tokenUsage, state, created)
+	def apply(text: String, thoughts: String = "", tokenUsage: TokenUsage = TokenUsage.zero,
+	          created: Instant = Now): BufferedResponse =
+		_BufferedResponse(text, thoughts, tokenUsage, created)
 	
 	
 	// NESTED   ------------------------
 	
-	private case class _BufferedResponse(text: String, tokenUsage: TokenUsage, state: Final, lastUpdated: Instant)
+	private case class _BufferedResponse(text: String, thoughts: String, tokenUsage: TokenUsage, lastUpdated: Instant)
 		extends BufferedResponse
 }
 
@@ -32,12 +31,17 @@ object BufferedResponse
  * @author Mikko Hilpinen
  * @since 02.09.2025, v1.4
  */
-trait BufferedResponse extends HasSchrodingerState
+trait BufferedResponse
 {
 	/**
 	 * @return Text contents of this response. Usually not including reflection / thinking content.
 	 */
 	def text: String
+	/**
+	 * @return The reflective / reasoning content produced by the LLM before the final answer.
+	 *         May be empty.
+	 */
+	def thoughts: String
 	/**
 	 * @return Time when the latest version of this response was originated.
 	 */
@@ -46,8 +50,4 @@ trait BufferedResponse extends HasSchrodingerState
 	 * @return Statistics about token usage in this interaction
 	 */
 	def tokenUsage: TokenUsage
-	/**
-	 * @return Whether this response was a success or a failure
-	 */
-	def state: Final
 }